test: 미션3 시간관리 API 테스트 코드 추가 #step4
